feat(dialogs): relocated dialogs in app
