+++ src/features/backlog-feature/backlog-feature.tsx
@@ -8,8 +8,8 @@ import { Button } from "@/components/ui/button/button";
 import { ButtonLoading } from "@/components/ui/button-loading/button-loading";
 import BacklogFeatureDetail from "@/features/backlog-feature/backlog-feature-detail/backlog-feature-detail";
 import BacklogFeatureList from "@/features/backlog-feature/backlog-feature-list/backlog-feature-list";
-import ConfigureRequirementsDialog from "@/features/refine-feature/configure-requirements-dialog/configure-requirements-dialog";
-import { RequirementTypeEnum } from "@/features/refine-feature/configure-requirements-dialog/configure-requirements-dialog.conts";
+import ConfigureRequirementsDialog from "@/features/shared/configure-requirements-dialog/configure-requirements-dialog";
+import { RequirementTypeEnum } from "@/features/shared/configure-requirements-dialog/configure-requirements-dialog.conts";
 import useRefineDetailBacklog from "@/hooks/use-refine-detail-backlog";
 import useRefineRequirements from "@/hooks/use-refine-requirements";
 import { useAppStore } from "@/stores/use-app-store";
