+++ src/features/refine-feature/refine-feature.tsx
@@ -4,11 +4,11 @@ import { useTranslation } from "react-i18next";
 import { Container } from "@/components/layout/container/container";
 import Sidebar from "@/components/sidebar/sidebar";
 import { Button } from "@/components/ui/button/button";
-import BackNewFeatureDialog from "@/features/refine-feature/back-new-feature-dialog/back-new-feature-dialog";
-import ConfigureRequirementsDialog from "@/features/refine-feature/configure-requirements-dialog/configure-requirements-dialog";
-import { RequirementTypeEnum } from "@/features/refine-feature/configure-requirements-dialog/configure-requirements-dialog.conts";
 import RefineFeatureChat from "@/features/refine-feature/refine-feature-chat/refine-feature-chat";
 import RefineFeatureRequirements from "@/features/refine-feature/refine-feature-requirements/refine-feature-requirements";
+import BackNewFeatureDialog from "@/features/shared/back-new-feature-dialog/back-new-feature-dialog";
+import ConfigureRequirementsDialog from "@/features/shared/configure-requirements-dialog/configure-requirements-dialog";
+import { RequirementTypeEnum } from "@/features/shared/configure-requirements-dialog/configure-requirements-dialog.conts";
 import { useAppStore } from "@/stores/use-app-store";
 
 const RefineFeature = () => {
